create select compoents for column and type
